% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABABmodel.R
\name{ABABmodel}
\alias{ABABmodel}
\title{Bayesian Estimation of Effect Sizes in Single-Case Experimental Designs}
\usage{
ABABmodel(y, P, s, model = "level", plots = TRUE)
}
\arguments{
\item{y}{outcome variable}

\item{P}{phase identifier}

\item{s}{session identifier}

\item{model}{the model to be fitted. It can be set to "level" (the default) or "trend". "level" model calculates intercepts only, and "trend" model calculates intercepts and slopes.}

\item{plots}{whether graphs are to be plotted. Defaults to TRUE.}
}
\value{
delta effect size estimates for A1B1, B1A2, and A2B2 phase changes
}
\description{
This function allows you to compute an interrupted time series (ITS) analysis with Bayesian estimates. It can be used in single-case experimental designs to examine the effect of the introduction of a treatment B after a baseline A and the retreat of such treatment thus the name of the model ABAB.
}
\examples{
library(SCEDbayes)
dat = data(LAMBERT, package = "SCEDbayes")
dat = subset(dat, dat$STUDENT==1)
model = ABABmodel(dat$DATA.POINT, dat$PHASE, dat$SESSION, model = 'level', plots = TRUE)
}
