% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABABmodel.R
\name{ABABmodel}
\alias{ABABmodel}
\title{Bayesian Estimation of Effect Sizes in ABAB Reversal Designs}
\usage{
ABABmodel(y, P, s, model = "level", plots = TRUE, diagnostics = FALSE)
}
\arguments{
\item{y}{outcome variable}

\item{P}{phase identifier}

\item{s}{session identifier}

\item{model}{the model to be fitted. If set to "level" (the default), the model calculates intercepts only. Set to "trend", the model calculates intercepts and slopes.}

\item{plots}{whether graphs are to be plotted. Defaults to TRUE.}
}
\value{
delta effect size estimates for A1B1, B1A2, and A2B2 phase changes
}
\description{
This function allows you to compute an interrupted time series (ITS) analysis with Bayesian estimates. It can be used with an ABAB reversal single-case experimental design to examine the effect of the introduction of a treatment B after a baseline A, followed up by a retreat and then reintroduction of such treatment.
}
\examples{
library(SCEDbayes)
dat = subset(LAMBERT, LAMBERT$STUDENT==1)
y = dat$DATA.POINT; P = dat$PHASE; s = dat$SESSION;
model1 = ABABmodel(y, P, s, model = 'level') # Intercepts only model
model2 = ABABmodel(y, P, s, model = 'trend') # Intercepts and slopes model
}
