% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABABmodel.R
\name{ABABmodel}
\alias{ABABmodel}
\title{Bayesian Estimation of Effect Sizes in Single-Case Experimental Designs}
\usage{
ABABmodel(y, P, s, model = "level")
}
\arguments{
\item{y}{outcome variable}

\item{P}{phase identifier
s session identifier
model the model to be fitted. It can be set to "level" (the default), "trend", or "AR1". "level" model calculates intercepts only, "trend" model calculates intercepts and slopes, and "AR1" calculates intercepts and an autocorrelation parameter at lag 1.}
}
\value{
Phase.estimates the level (and slopes if "trend" or rho if "AR1") bayesian estimates. Results include the posterior mean, median, and lower and apper 95% HDI.
Phase.change coefficient estimates for phase change at A1B1, B1A2, and A2B2. Results include the posterior mean, median, and lower and apper 95% HDI.
}
\description{
This function allows you to compute an interrupted time series (ITS) analysis with Bayesian estimates. It can be used in single-case experimental designs to examine the effect of the introduction of a treatment B after a baseline A and the retreat of such treatment thus the name of the model ABAB.
}
\examples{
library(SCEDbayes)
dat = data(LAMBERT, package = "SCEDbayes")
dat = subset(dat, dat$STUDENT==1)
model = ABABmodel(dat$DATA.POINT, dat$PHASE, dat$SESSION, model = 'level', plots = TRUE)
}
