% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABABmodel.R
\name{ABABmodel}

\alias{ABABmodel}

\title{Bayesian Estimation of Effect Sizes in Single-Case Experimental Designs}

\description{
This function computes an autoregression at lag 1 time series analysis with Bayesian estimates. It can be used in single-case experimental designs to examine the effect of the introduction of a treatment B after a baseline A and the retreat of such treatment, also called ABAB reversal designs.
}

\usage{
ABABmodel(y, P, s, model = "level", plots = TRUE)
}

\arguments{
\item{y}{outcome variable}
\item{P}{phase identifier}
\item{s}{session identifier}
\item{model}{the model to be fitted. If set to "level" (the default), the model calculates intercepts only. Set to "trend", the model calculates intercepts and slopes.}
\item{plots}{whether graphs are to be plotted. Defaults to TRUE.}
}

\value{
\item{beta}{model coefficient estimates for A1, A1B1, B1A2, and A2B2 phase changes}
\item{phases}{intercept and slope estimates for A1, B1, A2, and B2 phases}
\item{delta}{effect size estimates for A1B1, B1A2, and A2B2 phase changes}

}
\examples{
library(SCEDbayes)
dat = data(LAMBERT, package = "SCEDbayes")
dat = subset(dat, dat$STUDENT==1)
y = dat$DATA.POINT; P = dat$PHASE; s = dat$SESSION;
model.1 = ABABmodel(y, P, s, model = 'level')
model.2 = ABABmodel(y, P, s, model = 'trend')
}
