% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBmodel.R
\name{MBmodel}
\alias{MBmodel}
\title{Bayesian Estimation of Effect Sizes in Multiple-Baseline Designs}
\usage{
MBmodel(y, P, s, model = "level", plots = TRUE, diagnostics = FALSE)
}
\arguments{
\item{y}{outcome variable}

\item{P}{phase identifier}

\item{s}{session identifier}

\item{model}{the model to be fitted. If set to "level" (the default), the model calculates intercepts only. Set to "trend", the model calculates intercepts and slopes.}

\item{plots}{whether graphs are to be plotted. Defaults to TRUE.}
}
\value{
delta effect size estimates for A1B1, B1A2, and A2B2 phase changes
}
\description{
This function allows you to compute an interrupted time series (ITS) analysis with Bayesian estimates. It can be used in single-case experimental designs with multiple-baseline across participants to examine the effect of the introduction of a treatment B after a baseline A.
}
\examples{
library(SCEDbayes)
dat = subset(LAMBERT_MB, LAMBERT_MB$STUDENT==1)
y = dat$DATA.POINT; P = dat$PHASE; s = dat$SESSION;
model1 = MBmodel(y, P, s, model = 'level') # Intercepts only model
model2 = MBmodel(y, P, s, model = 'trend') # Intercepts and slopes model
}
