% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBmodel.R
\name{MBmodel}

\alias{MBmodel}

\title{Bayesian Estimation of Effect Sizes in Multiple-Baseline Designs}

\usage{
MBmodel(y, P, s, model = 'level', plots = TRUE, diagnostics = FALSE,
          adaptSteps = 10000, burnInSteps = 100000, nChains = 3,
            numSavedSteps = 200000, thinSteps = 10)
}

\arguments{
\item{y}{outcome variable}
\item{P}{phase identifier}
\item{s}{session identifier}
\item{model}{the model to be fitted. If set to "level" (the default), an intercepts only model is calculated. If set to "trend", an intercepts and slopes model is calculated.}
\item{plots}{whether graphs are to be plotted. Defaults to TRUE.}
\item{diagnostics}{whether diagnostic statistics are to be calculated. Defaults to FALSE. The diagnostics option retrieves the Gelman-Rubin statistic (which measures the ratio of the between-within variability within the chains); the ESS or effective sample size (which measures the independent information within the sampled chains given their length and autocorrelation); and the MCSE or Monte Carlo Standard Error (which measures the estimated standard deviation of the sample mean in the chains).}
\item{adaptSteps}{number of steps to adapt the chain}
\item{burnInSteps}{number of steps to burn in}
\item{nChains}{number of chains to be computed}
\item{numSavedSteps}{total number of steps to be computed}
\item{thinSteps}{save every nth number of steps in the chain}
}

\value{
\item{beta}{model regression coefficients estimates}
\item{phase}{phase regression coefficients estimates}
\item{delta}{phase change effect size estimates}
}

\description{
This function computes a lag-1 autoregression time series analysis with Bayesian estimates for multiple-baseline designs. It can be used with data from a multiple-baseline design to examine the effect of the introduction of a treatment B after a baseline A.
}

\examples{
library(SCEDbayes)
dat = subset(LAMBERT_MB, LAMBERT_MB$STUDENT==1)
y = dat$DATA.POINT; P = dat$PHASE; s = dat$SESSION;
model1 = MBmodel(y, P, s, model = 'level') # Intercepts only model
model2 = MBmodel(y, P, s, model = 'trend') # Intercepts and slopes model
}

\references{
Swaminathan, H., Rogers, H. J., & Horner, R. H. (2014). An effect size measure and Bayesian analysis of single-case designs. Journal of School Psychology, 52(2), 213-230. doi:10.1016/j.jsp.2013.12.002

Kruschke, J. (2010). Doing Bayesian Data Analysis: A Tutorial Introduction with R. Boston, MA.: Academic Press.
}
